fix: guard against errant uncaught exception
During startup, it's possible for the RedisClient to emit an error prior to emitting a ready event.
When we reject the Promise, the called doesn't receive a handle to the RedisClient, so any
subsequent errors will result in uncaught exceptions. It's also better practice to clean up that
client anyway.