fix(types): update redis option type (#290)
The Redis QueueSettings prop can accept a node_redis client object, but this wasn't reflected in the typings. If a Typescript user tried to set a RedisClient, Typescript would not compile without using `@ts-ignore`.

This PR fixes that by adding the `RedisClient` type to the `redis` prop as a union type.